requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
pillow==10.4.0
psycopg2-binary==2.9.9
sqlparse==0.5.1
tzdata==2024.2
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
coverage==7.6.3
django-cors-headers==4.5.0
drf-yasg==1.21.7
setuptools==75.1.0
